refactor: organize UserInfo folder files
